Worked on intermediary currency: - added a mock for getting intermediary currency code - added code for fetching exchange rate using intermediary currency - refactored

+++ app/exchange_rates/schemes/exchange_rates.py
@@ -22,7 +22,7 @@ class ExchangeRateRequest(BaseModel):
     currency_from: str
     currency_to: str
     exchange: ExchangeEnum | None | str
-    amount: int
+    amount: Decimal
     cache_max_seconds: int | None
